requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.5
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pendulum==3.1.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
SQLAlchemy==2.0.42
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
reportlab==4.2.5
Pillow==10.4.0
